+++ plugins/modules/oci_data_flow_run.py
@@ -110,6 +110,10 @@ options:
             - An Oracle Cloud Infrastructure URI of the bucket where the Spark job logs are to be uploaded.
               See https://docs.cloud.oracle.com/iaas/Content/API/SDKDocs/hdfsconnector.htm#uriformat.
         type: str
+    metastore_id:
+        description:
+            - The OCID of OCI Hive Metastore.
+        type: str
     num_executors:
         description:
             - The number of executor VMs requested.
@@ -342,6 +346,12 @@ run:
             returned: on success
             type: string
             sample: logs_bucket_uri_example
+        metastore_id:
+            description:
+                - The OCID of OCI Hive Metastore.
+            returned: on success
+            type: string
+            sample: "ocid1.metastore.oc1..xxxxxxEXAMPLExxxxxx"
         num_executors:
             description:
                 - The number of executor VMs requested.
@@ -488,6 +498,7 @@ run:
         "lifecycle_details": "lifecycle_details_example",
         "lifecycle_state": "ACCEPTED",
         "logs_bucket_uri": "logs_bucket_uri_example",
+        "metastore_id": "ocid1.metastore.oc1..xxxxxxEXAMPLExxxxxx",
         "num_executors": 56,
         "opc_request_id": "ocid1.opcrequest.oc1..xxxxxxEXAMPLExxxxxx",
         "owner_principal_id": "ocid1.ownerprincipal.oc1..xxxxxxEXAMPLExxxxxx",
@@ -644,6 +655,7 @@ def main():
             executor_shape=dict(type="str"),
             freeform_tags=dict(type="dict"),
             logs_bucket_uri=dict(type="str"),
+            metastore_id=dict(type="str"),
             num_executors=dict(type="int"),
             parameters=dict(
                 type="list",