+++ plugins/modules/oci_database_autonomous_database_clones_facts.py
@@ -767,6 +767,15 @@ autonomous_database_clones:
             returned: on success
             type: list
             sample: []
+        autonomous_maintenance_schedule_type:
+            description:
+                - The maintenance schedule type of the Autonomous Database on shared Exadata infrastructure. The EARLY maintenance schedule of this Autonomous
+                  Database
+                  follows a schedule that applies patches prior to the REGULAR schedule.The REGULAR maintenance schedule of this Autonomous Database follows the
+                  normal cycle.
+            returned: on success
+            type: string
+            sample: EARLY
     sample: [{
         "id": "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx",
         "compartment_id": "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx",
@@ -867,7 +876,8 @@ autonomous_database_clones:
         "time_local_data_guard_enabled": "2013-10-20T19:20:30+01:00",
         "dataguard_region_type": "PRIMARY_DG_REGION",
         "time_data_guard_role_changed": "2013-10-20T19:20:30+01:00",
-        "peer_db_ids": []
+        "peer_db_ids": [],
+        "autonomous_maintenance_schedule_type": "EARLY"
     }]
 """