+++ plugins/modules/oci_golden_gate_deployment_backup.py
@@ -297,6 +297,9 @@ except ImportError:
 class DeploymentBackupHelperGen(OCIResourceHelperBase):
     """Supported operations: create, update, get, list and delete"""
 
+    def get_entity_type(self):
+        return "goldengatedeploymentbackup"
+
     def get_module_resource_id_param(self):
         return "deployment_backup_id"