+++ plugins/modules/oci_mysql_backup.py
@@ -24,6 +24,7 @@ short_description: Manage a Backup resource in Oracle Cloud Infrastructure
 description:
     - This module allows the user to create, update and delete a Backup resource in Oracle Cloud Infrastructure
     - For I(state=present), create a backup of a DB System.
+    - "This resource has the following action operations in the M(oci_backup_actions) module: change_compartment."
 version_added: "2.9"
 author: Oracle (@oracle)
 options: